Perf Tests: Wild attempt at improving test determinism
In this patch we're tossing in Chrome CLI args in an attempt to run things in a
more deterministic manner, as well as trying to improve the timer precision, which
is normally reduced as a mitigation against speculative execution attacks.

The goal is to be able to run the performance tests CI workflow on a branch against
itself or against another branch with a no-code change (such as a Markdown doc update)
and end up with performance test results that are close enough to each other to be
effectively equal.

Currently the tests run with variation in the results that exceeds any actual variation
between the branches, whereas we find statistical confidence that random noise is unlikely
to account for the differences in the readings that we measure.

Hopefully we can adjust some command-line arguments and figure out that some of them
will help with the test reliability and we can add those to the repository.

https://peter.sh/experiments/chromium-command-line-switches/
https://chromium.googlesource.com/v8/v8/+/master/src/flags/flag-definitions.h#188