Tag Processor: Only define classes if they don't already exist.
In preparation for a merge into WordPress 6.2 we're wrapping the
class definitions for the HTML Tag Processor and its dependent
helper classes in checks for whether they already exist.

Additionaly we've renamed the `index.php` file to `wp-html.php`
to better align with its role of loading the WordPress HTML
parsing subsystem, both with the Tag Processor itself but also
with future higher-level APIs and expansions on top of it.

Co-Authored-By: Bernie Reiter <ockham@raz.or.at>